+++ CHANGELOG.md
@@ -16,11 +16,14 @@ The versioning scheme is listed in the README.
 
 ## Unreleased - DATE
 
+### Changed
+
+- Sprint/sneak heights can be smaller/larger than the configured step height.
+
 ### Fixed
 
 - Sprint height is only used when holding forward.
   - This prevents a bug where the sprint height was being used while moving sideways or backwards when toggle sprint is enabled.
-
 
 ## v2.3.1 - 2026-05-17